requirements.txt
Django==4.1.5
rest_framework==3.14.0
rest_framework_simplejwt==5.2.2
Pillow==9.4.0
gunicorn==20.1.0